Increased prime performance about 3x
The composites list had a lot of even numbers;
since even numbers are already skipped,
we don't need to include any of those composites.
Easy enough to do, just remember to square new ps
initially, and to add them twice rather than once
when updating.
(squaring initially is okay because the primes
 beneath it will hit the odd multiples with the
 new prime before the square is reached)

Starting to wonder if there is cleverness that
we could work in with 3 as well. Saving that for
a future date, I suppose.